+++ src/client/java/dev/isxander/yacl/gui/RequireRestartScreen.java
