+++ src/client/java/dev/isxander/yacl/gui/controllers/cycling/EnumController.java
