+++ src/client/java/dev/isxander/yacl/gui/controllers/slider/DoubleSliderController.java
