+++ src/client/java/dev/isxander/yacl/impl/utils/DimensionIntegerImpl.java
