+++ src/client/resources/yet-another-config-lib.client.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "dev.isxander.yacl.mixin",
+  "package": "dev.isxander.yacl.mixin.client",
   "compatibilityLevel": "JAVA_17",
   "injectors": {
     "defaultRequire": 1
