+++ src/main/java/dev/isxander/yacl/config/ConfigInstance.java
@@ -1,9 +1,6 @@
 package dev.isxander.yacl.config;
 
-import dev.isxander.yacl.api.YetAnotherConfigLib;
-
 import java.lang.reflect.InvocationTargetException;
-import java.util.function.BiFunction;
 
 /**
  * Responsible for handing the actual config data type.
@@ -44,10 +41,6 @@ public abstract class ConfigInstance<T> {
 
     public Class<T> getConfigClass() {
         return this.configClass;
-    }
-
-    public YetAnotherConfigLib buildConfig(BiFunction<ConfigInstance<T>, YetAnotherConfigLib.Builder, YetAnotherConfigLib.Builder> builder) {
-        return builder.apply(this, YetAnotherConfigLib.createBuilder().save(this::save)).build();
     }
 
     public abstract void save();
